refactor: Company 관련 코드 패키지 이동 (#301)
