+++ src/main/java/com/soda/global/init/TestDataLoader.java
@@ -1,6 +1,6 @@
 //package com.soda.global.init;
 //
-//import com.soda.member.domain.Company;
+//import com.soda.member.domain.company.Company;
 //import com.soda.member.domain.Member;
 //import com.soda.project.domain.company.enums.CompanyProjectRole;
 //import com.soda.project.domain.member.enums.MemberProjectRole;
