+++ src/main/java/com/soda/member/domain/Member.java
@@ -1,6 +1,7 @@
 package com.soda.member.domain;
 
 import com.soda.common.BaseEntity;
+import com.soda.member.domain.company.Company;
 import com.soda.notification.entity.MemberNotification;
 import com.soda.project.domain.member.MemberProject;
 import jakarta.persistence.*;
