+++ src/main/java/com/soda/member/domain/company/CompanyErrorCode.java
@@ -1,4 +1,4 @@
-package com.soda.member.domain;
+package com.soda.member.domain.company;
 
 import com.soda.global.response.ErrorCode;
 import org.springframework.http.HttpStatus;
