+++ src/main/java/com/soda/member/infrastructure/CompanyRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.soda.member.infrastructure;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.soda.member.domain.QCompany;
+import com.soda.member.domain.company.QCompany;
 import com.soda.member.interfaces.dto.CompanyCreationStatRaw;
 import lombok.RequiredArgsConstructor;
