+++ src/main/java/com/soda/project/application/stage/article/vote/validator/VoteValidator.java
@@ -1,7 +1,7 @@
 package com.soda.project.application.stage.article.vote.validator;
 
 import com.soda.global.response.GeneralException;
-import com.soda.member.domain.Company;
+import com.soda.member.domain.company.Company;
 import com.soda.member.domain.Member;
 import com.soda.member.domain.MemberRole;
 import com.soda.project.domain.Project;
