+++ src/main/java/com/soda/project/domain/company/CompanyProject.java
@@ -1,7 +1,7 @@
 package com.soda.project.domain.company;
 
 import com.soda.common.BaseEntity;
-import com.soda.member.domain.Company;
+import com.soda.member.domain.company.Company;
 import com.soda.project.domain.company.enums.CompanyProjectRole;
 import com.soda.project.domain.Project;
 import jakarta.persistence.*;
