+++ src/main/java/com/soda/project/infrastructure/CompanyProjectRepository.java
@@ -1,6 +1,6 @@
 package com.soda.project.infrastructure;
 
-import com.soda.member.domain.Company;
+import com.soda.member.domain.company.Company;
 import com.soda.project.domain.company.enums.CompanyProjectRole;
 import com.soda.project.domain.company.CompanyProject;
 import com.soda.project.domain.Project;
