+++ src/main/java/com/soda/project/interfaces/dto/ProjectCreateResponse.java
@@ -1,6 +1,6 @@
 package com.soda.project.interfaces.dto;
 
-import com.soda.member.domain.Company; // Company import
+import com.soda.member.domain.company.Company; // Company import
 import com.soda.member.domain.Member;   // Member import
 import com.soda.project.domain.member.enums.MemberProjectRole; // 역할 확인용 (선택적)
 import com.soda.project.domain.Project;
